[COMPLETE] 16-5

+++ src/Chapter16/Problem5.java
@@ -14,7 +14,7 @@ public class Problem5 {
         System.out.println(p5.findLastZeros(p5.factorialN(s.nextInt())));
     }
 
-    public int findLastZeros(int num){
+    public int findLastZeros(long num){
         int cnt = 0;
         while(num%10==0){
             num/=10;
@@ -23,8 +23,8 @@ public class Problem5 {
         return cnt;
     }
 
-    public int factorialN(int n){
-        int result=1;
+    public long factorialN(int n){
+        long result=1;
         for(int i=2;i<=n;i++){
             result*=i;
         }
